Network.HTTP.Link.Parser

Description

The parser for the HTTP Link header as defined in RFC 5988. More liberal than the RFC though: does not validate URLs and other deep stuff, accepts whitespace in weird places.

Documentation

linkHeader :: IsURI uri => Parser [Link uri]

The Attoparsec parser for the Link header.

parseLinkHeader' :: IsURI uri => Text -> Either String [Link uri]

Parses a Link header, returns an Either, where Left is the Attoparsec error string (probably not a useful one).

parseLinkHeader :: IsURI uri => Text -> Maybe [Link uri]

Parses a Link header, returns a Maybe.

parseLinkHeaderBS' :: IsURI uri => ByteString -> Either String [Link uri]

Parses a Link header, returns an Either, where Left is the Attoparsec error string (probably not a useful one).

parseLinkHeaderBS :: IsURI uri => ByteString -> Maybe [Link uri]

Parses a Link header, returns a Maybe.