Network.HTTP.Link.Parser

Description

The parser for the HTTP Link header as defined in RFC 5988. More liberal than the RFC though: does not validate URLs and other deep stuff, accepts whitespace in weird places.

Documentation

linkHeader :: Parser [Link]

The Attoparsec parser for the Link header.

parseLinkHeader' :: Text -> Either String [Link]

Parses a Link header, returns an Either, where Left is the Attoparsec error string (probably not a useful one).

parseLinkHeader :: Text -> Maybe [Link]

Parses a Link header, returns a Maybe.

parseLinkHeaderBS' :: ByteString -> Either String [Link]

Parses a Link header, returns an Either, where Left is the Attoparsec error string (probably not a useful one).

parseLinkHeaderBS :: ByteString -> Maybe [Link]

Parses a Link header, returns a Maybe.